modify time_and_future.cpp

+++ time_and_future.cpp
@@ -77,8 +77,8 @@ void sysLocalTime(string info)
         cout.fill('.');
 
         cout << left << setw(30) << info;
-        printf(" %04d-%02d-%02d %02d:%02d:%02d.%03d\n", 
-                        1900+p->tm_year, 1+p->tm_mon, p->tm_mday, p->tm_hour, p->tm_min, p->tm_sec, int(double(tv.tv_usec)/1000));  
+        printf(" %04d-%02d-%02d %02d:%02d:%02d.%03ld\n", 
+                        1900+p->tm_year, 1+p->tm_mon, p->tm_mday, p->tm_hour, p->tm_min, p->tm_sec, tv.tv_usec/1000);  
 }
